# lib/api-zod/src/generated/types/statsResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * APEX Nifty 50 Trading Scanner API
 * OpenAPI spec version: 0.1.0
 */

export interface StatsResponse {
  total_symbols: number;
  active_trades: number;
  pending_signals: number;
  buy_signals: number;
  sell_signals: number;
  /** @nullable */
  last_scan?: string | null;
  scanning: boolean;
  scan_errors: number;
  /** Number of scan cycles that have actually started since service startup */
  scan_count: number;
  /**
     * Monotonic duration of the most recently completed scan cycle in milliseconds
     * @nullable
     */
  scan_latency_ms: number | null;
  /**
     * ISO timestamp when the current or most recent scan cycle started
     * @nullable
     */
  scan_started_at: string | null;
  timeframes: string[];
  /** OPEN | PRE_OPEN | CLOSED | HOLIDAY | WEEKEND */
  session_status: string;
  market_open: boolean;
  nse_time: string;
  /** @nullable */
  next_open?: string | null;
  /** @nullable */
  next_close?: string | null;
}